# docs/openapi3.yml
openapi: 3.0.0
...
servers:
  - description: localhost
    url: http://localhost:3002/api
info:
  description: Cloud TP Note
  version: '1.0.0'
  title: Cloud TP Note
  contact:
    email: nocompany@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
components:
  schemas:
    Task:
      type: object
      required:
        - id
        - title
        - priorite
      properties:
        id:
          type: string
          format: id
          example: 0beda6e7-874b-4319-8651-2e3aba504463
        title:
          type: string
          example: livre unique
        description:
          type: string
          example: unique et important
        priorite:
          type: integer
          example: 4
...
paths:
...
  /Task/{id}:
    get:
      operationId: GetTask
      description: il récupère une tache par son id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    patch:
      operationId: UpdateTask
      description: il met a jour une tache
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    delete:
      operationId: DelTask
      description: il supprime une tache par son id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
...